@@ -55,7 +55,7 @@ var fromBuffer = function(b) {
 
 var toBuffer = function() {
   var buffer, hex, i, key, octet, opt, padded, pos, value, _i, _j, _k, _l, _len, _len1, _len2, _len3, _ref, _ref1, _ref2, _ref3;
-  buffer = new Buffer(512, 'ascii');
+  buffer = Buffer.alloc(512, 0x00, 'ascii');
   buffer[0] = this.op;
   buffer[1] = this.htype;
   buffer.writeUInt8(this.hlen, 2);
@@ -96,7 +96,7 @@ var toBuffer = function() {
     }
   }
   buffer[pos] = 255;
-  padded = new Buffer(pos, 'ascii');
+  padded = Buffer.alloc(pos, 0x00, 'ascii');
   buffer.copy(padded, 0, 0, pos);
   return padded;
 };
